# More `convert` fixes, parameterize on flagtype