Ref: https://discourse.julialang.org/t/ann-sumtypes-jl-v0-4/97038/7

Before:
```julia
using SumTypes, BenchmarkTools

@sum_type Result{T} begin
  Failure
  Success{T}(::T)
end

# If I annotate the return type here as ::Result{T}, this function throws an
# ambiguous method error.
function log_nothrow(x::T) where{T<:AbstractFloat}
  if x < zero(x) 
    # Because I can't annotate the return type, using the other README trick.
    f::Result{T} = Failure
    return f
  end
  Success(log(x))
end

julia> @benchmark log_nothrow($(1.1))
BenchmarkTools.Trial: 10000 samples with 998 evaluations.
 Range (min … max):  15.360 ns … 25.671 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     15.371 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   15.394 ns ±  0.168 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▃    █         ▂                                             
  █▁▁▁▃█▁▁▁▂▃▁▁▁▂█▁▁▁▆▁▁▁▁▂▁▁▁▁▂▁▁▁▁▂▁▁▁▁▂▁▁▁▂▅▁▁▁▂▂▁▁▁▂▃▁▁▁▃ ▂
  15.4 ns         Histogram: frequency by time        15.5 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

After:
```julia
function log_nothrow(x::T)::Result{T} where{T<:AbstractFloat}
  if x < zero(x) 
      return Failure
  end
  Success(log(x))
end

julia> @benchmark log_nothrow(1.1)
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  1.080 ns … 4.470 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.090 ns             ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.090 ns ± 0.038 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                              █                              
  ▅▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂ ▂
  1.08 ns        Histogram: frequency by time        1.1 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
(the slowdown in the other way of writing it was also fixed):
```julia
julia> function log_nothrow(x::T) where{T<:AbstractFloat}
         if x < zero(x) 
           # Because I can't annotate the return type, using the other README trick.
           f::Result{T} = Failure
           return f
         end
         Success(log(x))
       end
log_nothrow (generic function with 1 method)

julia> @benchmark log_nothrow(1.1)
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  1.079 ns … 2.770 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.090 ns             ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.088 ns ± 0.034 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

       ▂                                                  █  
  ▂▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▁▁▁▁█ ▂
  1.08 ns        Histogram: frequency by time       1.09 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
_________________________

After adding some convert methods, there's still a performance loss due to alignment difficulties:
Before
```julia
julia> let x = Ref(1.1)
           @btime log_nothrow($x[])
           @btime log($x[])
       end
  15.280 ns (0 allocations: 0 bytes)
  4.740 ns (0 allocations: 0 bytes)
0.09531017980432493
```

I've now changed things so that the size of the discriminator byte actually changes to ensure that julia makes good alignment choices. This required the sum type to be parameterized on the flagtype as well, which is unfortunate but does work:
```julia
julia> let x = Ref(1.1)
           @btime log_nothrow($x[])
           @btime log($x[])
       end
  5.220 ns (0 allocations: 0 bytes)
  4.740 ns (0 allocations: 0 bytes)
0.09531017980432493
```